formalized the InputTest with the InputLobby script.
