+++ Assets/Scripts/Characters/Player/Input/InputLobby.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6961c19c9a2551a03bdd0618da5560d2
+guid: 9ecf74fb409a0f36393352baecfcf28d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
